Fixing pull request problems
