+++ git-test/src/components/Home/Home.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { Button, Card, CardContent, CardMedia, Typography, CardHeader, Box } from '@mui/material';
 import ReactQuill from 'react-quill';
 import 'react-quill/dist/quill.snow.css'; 
-import { PokemonData } from './interfaces/PokemonInterfaces';
-import fetchPokemonData from './fetchPokemonData';
+import { PokemonData } from '../../interfaces/PokemonInterfaces';
+import { fetchPokemonData } from '../../fetchPokemonData';
 import { homeStyles } from './homeStyles';
