+++ git-test/src/components/Home/homeStyles.ts
@@ -49,3 +49,4 @@ export const homeStyles = makeStyles({
       borderRadius: '8px',
     },
   });
+
