+++ git-test/src/fetchPokemonData.ts
@@ -1,10 +1,9 @@
 import { PokemonData } from './interfaces/PokemonInterfaces';
 
-const fetchPokemonData = async (): Promise<PokemonData> => {
+export const fetchPokemonData = async (): Promise<PokemonData> => {
   const endpointUrl = import.meta.env.VITE_ENDPOINT_URL;
   const response = await fetch(endpointUrl);
   const data: PokemonData = await response.json();
   return data;
 };
 
-export default fetchPokemonData;
