+++ git-test/src/interfaces/PokemonInterfaces.ts
