+++ git-test/src/main.tsx
@@ -1,10 +1,15 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
+import theme from './colors/commonTheme.ts'
 import './index.css'
+import { ThemeProvider,CssBaseline } from '@mui/material'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <App />
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+        <App />
+    </ThemeProvider>
   </React.StrictMode>,
 )
